Added support to HostapdConfig POST to manager

+++ datamodel/datamodel.go
@@ -11,7 +11,7 @@ type ProcessPid struct {
 // interface
 type HostapdConfig struct {
 	Interface     string `json:"interface, omitempty"`
-	ReauthTimeout int    `json:"reauth_timeout, omitempty"`
+	ReauthTimeout uint64 `json:"reauth_timeout, omitempty"`
 	OpenvSwitch   string `json:"openvswitch, omitempty"`
 }
